Removed RemoveValueAt (...)
RemoveValueAt doesn't tell you if it fails because there is no value or
because the value that's there doesn't match. An option to fix it is to
add an extra out parameter, another option to fix it is to loop. I have
decided to remove it because I don't think it is worth to keep it.