Implemented "Combos"
TryGetCheckRemoveAt will try to retrive a value, if got it then check it
and remove it.
TryGetChekSet will try to retrieve a value, if got it then check it and
replace it, otherwise insert.

The value may have been replaced during check by another thread, this is
ignored... and that is why these methods are Internal.